requirements.txt
fastapi
uvicorn
python-dotenv
openai
PyPDF2
python-docx
openpyxl